M: fix bug in nodeserver/service.go resp should SetReturnNum; fix bug in Client about int&int8; unify the client/server called function name and parameter type; fix several bugs in Server; add ResolveTCPAddr err test; build Client&NodeServer communication directly

+++ Client/jgrpc/interface.go
@@ -11,6 +11,6 @@ type Calculator struct {
 }
 
 // Add is a function to calculate a + b
-func (c *Calculator) Add(a, b int) int {
+func (c *Calculator) Add(a, b int8) int8 {
 	return add(c, a, b)
 }
